✅🤖 organize download tests around their contracts
